+++ RELEASES.md
@@ -1,3 +1,15 @@
+Carrot2-v9 (2025-12-06)
+========================
+* DarkSouls model
+* fix Angle Steering(HKG car)
+* fix LaneChange desire
+
+Carrot2-v9 (2025-12-06)
+========================
+* PP(planplus) model
+* fixDM
+* fix angle steering torque(HKG car)
+
 Carrot2-v9 (2025-12-03)
 ========================
 * ST model